+++ services/src/main/java/com/allsoftdroid/audiobook/services/audio/broadcastReceivers/NotificationPlayerEventBroadcastReceiver.kt
@@ -1,4 +1,4 @@
-package com.allsoftdroid.audiobook.services.audio
+package com.allsoftdroid.audiobook.services.audio.broadcastReceivers
 
 import android.app.PendingIntent
 import android.content.BroadcastReceiver
